wheat-backend-main/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
gdown
python-multipart

